refactor(extension): rename ControlAndPluginService to ExtensionService
